+++ iotdb/include/Catalogs/PhysicalStreamConfig.hpp
@@ -19,7 +19,7 @@ struct PhysicalStreamConfig {
   PhysicalStreamConfig();
 
   PhysicalStreamConfig(std::string sourceType, std::string sourceConfig,
-                       double sourceFrequency, size_t numberOfBuffersToProduce,
+                       size_t sourceFrequency, size_t numberOfBuffersToProduce,
                        std::string physicalStreamName,
                        std::string logicalStreamName);
 
@@ -27,7 +27,7 @@ struct PhysicalStreamConfig {
 
   std::string sourceType;
   std::string sourceConfig;
-  double sourceFrequency;
+  size_t sourceFrequency;
   size_t numberOfBuffersToProduce;
   std::string physicalStreamName;
   std::string logicalStreamName;
